// dotnet/openapi/model/enquiryModelListPagedResponse.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EnquiryModel } from './enquiryModel';


export interface EnquiryModelListPagedResponse { 
    message?: string | null;
    statusCode?: number;
    data?: Array<EnquiryModel> | null;
    pageIndex?: number;
    pageSize?: number;
    totalRecords?: number;
}