// dotnet/openapi/model/getAllEnquiriesQuery.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetAllEnquiriesQuery { 
    pageIndex?: number;
    pageSize?: number;
    searchString?: string | null;
}